+++ dist_app/SpritesMng.js
@@ -1,5 +1,5 @@
 import { U as w, L as v, V as C, S as L, m as h, i as x, y as R, D as N, b as F, N as A, B as k, x as y, e as H, z as Y, k as O } from "./app2.js";
-import { a as V } from "./Reading.js";
+import { a as V } from "./CmnTween.js";
 class e {
   constructor(t = "", a, c = () => {
   }, n = () => {