+++ dist_app/Variable.js
@@ -1,7 +1,7 @@
 import { u as b, f as p, h as w, i as f, j as N, k as r, e as S } from "./app2.js";
 import { c as O } from "./CallStack.js";
 import { PropParser as C } from "./PropParser.js";
-import { R as W } from "./Reading.js";
+import { R as W } from "./CmnTween.js";
 class m {
   #t = {};
   clear() {