+++ dist_app/sn/Reading.d.ts
@@ -16,6 +16,7 @@ type IPageLog = {
 };
 export declare class ReadingState {
     #private;
+    static init(): void;
     static get rs(): ReadingState;
     constructor();
     static setEvt2Fnc(glb: boolean, key: string, fnc: T_Evt2Fnc): void;