+++ src/sn/Pages.ts
@@ -13,6 +13,7 @@ import {argChk_Boolean} from './CmnLib';
 
 import type {Container} from 'pixi.js';
 
+
 export class Pages {
 	#pg: {fore: Layer, back: Layer};