+++ src/sn/RubySpliter.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-non-null-assertion */
 /* eslint-disable no-irregular-whitespace */
 /* ***** BEGIN LICENSE BLOCK *****
 	Copyright (c) 2018-2025 Famibee (famibee.blog38.fc2.com)
@@ -25,14 +24,14 @@ export class RubySpliter {
 
 /*
 		★Unicodeで「漢字」の正規表現 – ものかの http://tama-san.com/kanji-regex/
-		2E80..2FDF CJK部首補助＋康熙部首
-		3005 々（漢字の踊り字）
-		3007 〇（漢数字のゼロ）
-		303B 〻（漢字の踊り字）
-		3400..4DBF CJK統合漢字拡張A
-		4E00..9FFF CJK統合漢字
-		F900..FAFF CJK互換漢字
-		20000..2FFFF CJK統合漢字拡張B〜F＋CJK互換漢字追加＋念のためU+2FFFFまで
+		2E80..2FDF	CJK部首補助＋康熙部首
+		3005		々（漢字の踊り字）
+		3007		〇（漢数字のゼロ）
+		303B		〻（漢字の踊り字）
+		3400..4DBF		CJK統合漢字拡張A
+		4E00..9FFF		CJK統合漢字
+		F900..FAFF		CJK互換漢字
+		20000..2FFFF	CJK統合漢字拡張B〜F＋CJK互換漢字追加＋念のためU+2FFFFまで
 
 		[\x{2E80}-\x{2FDF}々〇〻\x{3400}-\x{4DBF}\x{4E00}-\x{9FFF}\x{F900}-\x{FAFF}\x{20000}-\x{2FFFF}]
 		[\u2E80-\u2FDF々〇〻\u3400-\u4DBF\u4E00-\u9FFF\uF900-\uFAFF\u20000-\u2FFFF]
@@ -75,7 +74,8 @@ export class RubySpliter {
 
 	putTxt(text: string) {
 		for (const {groups} of text.matchAll(RubySpliter.#REG_RUBY)) {
-			const {ruby, kan_ruby, kan='', ce, txt='', str=''} = groups!;
+			if (! groups) break;
+			const {ruby, kan_ruby, kan='', ce, txt='', str=''} = groups;
 			if (ruby) {this.putTxtRb(decodeURIComponent(str), ruby); continue}
 
 			if (kan_ruby) {this.putTxtRb(kan, kan_ruby); continue}
@@ -94,7 +94,7 @@ export class RubySpliter {
 		const len = a.length;
 		if (/^\*.?$/.test(ruby)) {	// 傍点文法
 			const rb_ses = 'center｜'+ (ruby === '*' ?RubySpliter.#sesame :ruby.charAt(1));
-			for (let i=0; i<len; ++i) this.#putCh(a[i]!, rb_ses);
+			for (const s of a) this.#putCh(s, rb_ses);
 			return;
 		}
 
@@ -109,7 +109,9 @@ export class RubySpliter {
 		const lenR = aR.length;
 		const len_max = lenR > len ?lenR :len;
 		for (let i=0; i<len_max; ++i) this.#putCh(
+			// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 			i < len ? a[i]! : '',
+			// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 			i < lenR ? decodeURIComponent(aR[i]!) : ''
 		);
 	}